requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
requests==2.25.1
smmap==3.0.5
soupsieve==2.2
urllib3==1.26.4
